# evidenciesib-api-externa/openapi/openapi_comanda.json
{
  "openapi" : "3.0.1",
  "servers" : [ {
    "url" : "/evidenciesibapi/externa"
  }, {
    "url" : "http://localhost:8080/evidenciesibapi/externa"
  }, {
    "url" : "https://dev.caib.es/evidenciesibapi/externa"
  }, {
    "url" : "https://proves.caib.es/evidenciesibapi/externa"
  }, {
    "url" : "https://se.caib.es/evidenciesibapi/externa"
  }, {
    "url" : "https://www.caib.es/evidenciesibapi/externa"
  } ],
  "paths" : {
...
    "/secure/logs/v1/{nomFitxer}/directe" : {
      "get" : {
        "operationId" : "descarregarFitxerDirecte",
        "parameters" : [ {
          "name" : "nomFitxer",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/octet-stream" : { }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/logs/v1/{nomFitxer}" : {
      "get" : {
        "operationId" : "getFitxerByNom",
        "parameters" : [ {
          "name" : "nomFitxer",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FitxerContingut"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/logs/v1/{nomFitxer}/linies/{nLinies}" : {
      "get" : {
        "operationId" : "llegitUltimesLinies",
        "parameters" : [ {
          "name" : "nomFitxer",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "nLinies",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/logs/v1" : {
      "get" : {
        "operationId" : "llistarFitxers",
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FitxerInfo"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "FitxerContingut" : {
        "type" : "object",
        "properties" : {
          "nom" : {
            "type" : "string"
          },
          "mida" : {
            "type" : "integer",
            "format" : "int64"
          },
          "mimeType" : {
            "type" : "string"
          },
          "dataCreacio" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dataModificacio" : {
            "type" : "string",
            "format" : "date-time"
          },
          "contingut" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "byte"
            }
          }
        }
      },
      "FitxerInfo" : {
        "type" : "object",
        "properties" : {
          "nom" : {
            "type" : "string"
          },
          "mida" : {
            "type" : "integer",
            "format" : "int64"
          },
          "mimeType" : {
            "type" : "string"
          },
          "dataCreacio" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dataModificacio" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      }
    }
  }
}