# evidenciesib-api-externa/src/main/java/es/caib/evidenciesib/api/externa/secure/comanda/v1/ComandaLogService.java
package es.caib.evidenciesib.api.externa.secure.comanda.v1;

import java.util.List;

import javax.annotation.security.RolesAllowed;
import javax.ws.rs.GET;
import javax.ws.rs.Path;
import javax.ws.rs.PathParam;
import javax.ws.rs.Produces;

import org.fundaciobit.pluginsib.utils.rest.RestUtils;

import es.caib.comanda.model.server.monitoring.FitxerContingut;
import es.caib.comanda.model.server.monitoring.FitxerInfo;
import es.caib.comanda.ms.log.helper.LogFileStream;
import es.caib.comanda.ms.log.helper.LogHelper;
import es.caib.evidenciesib.commons.utils.Constants;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.security.SecurityScheme;

import javax.ws.rs.core.Response;
import java.io.InputStream;

/**
 * 
 * @author anadal
 * 4 feb 2026 12:45:47
 */
@Path("/secure/logs/v1")
@io.swagger.annotations.SwaggerDefinition(
        info = @io.swagger.annotations.Info(
                title = "ComandaAppLogs API",
                version = "1.0",
                description = "API per a la gestió de logs de l'aplicació Comanda"),
        tags = { @io.swagger.annotations.Tag(
                name = "COMANDA → APP / Logs",
                description = "Operacions relacionades amb els fitxers de log de l'aplicació Comanda") })
//@Api(description = "the ComandaAppLogs API")
@SecurityScheme(type = SecuritySchemeType.HTTP, name = ComandaServerUtils.SECURITY_NAME, scheme = "basic")
public class ComandaLogService extends RestUtils implements es.caib.comanda.api.server.monitoring.ComandaAppLogsApi {

    /**
     * Retorna el contingut i detalls del fitxer de log que es troba dins la carpeta de logs del servidor, i que té el nom indicat
     *
     * @param nomFitxer Nom del fitxer
     * @return successful operation
     */
    @GET
    @Path("/{nomFitxer}")
    @Produces({ "application/json" })
    @ApiOperation(
            value = "Obtenir contingut complet d'un fitxer de log",
            notes = "Retorna el contingut i detalls del fitxer de log que es troba dins la carpeta de logs del servidor, i que té el nom indicat",
            tags = { "COMANDA → APP / Logs" })
    @ApiResponses(
            value = { @ApiResponse(code = 200, message = "successful operation", response = FitxerContingut.class) })
    @RolesAllowed({ Constants.EVI_WS })
    @SecurityRequirement(name = ComandaServerUtils.SECURITY_NAME)
    @Override
    public FitxerContingut getFitxerByNom(@PathParam("nomFitxer") @ApiParam("Nom del firxer")
    String nomFitxer) {

        return LogHelper.getFitxerByNom(LogHelper.getDirectoryLogsFromJbossServerProperties(), nomFitxer);

        //return ComandaServerUtils.getFitxerByNom(nomFitxer);
    }

    /**
     * Retorna les darreres línies del fitxer de log indicat per nom. Concretament es retorna el número de línies indicat al paràmetre nLinies.
     *
     * @param nomFitxer Nom del firxer
     * @param nLinies Número de línies a recuperar del firxer
     * @return successful operation
     */
    @GET
    @Path("/{nomFitxer}/linies/{nLinies}")
    @Produces({ "application/json" })
    @ApiOperation(
            value = "Obtenir les darreres línies d'un fitxer de log",
            notes = "Retorna les darreres linies del fitxer de log indicat per nom. Concretament es retorna el número de línies indicat al paràmetre nLinies.",
            tags = { "COMANDA → APP / Logs" })
    @ApiResponses(
            value = { @ApiResponse(
                    code = 200,
                    message = "successful operation",
                    response = String.class,
                    responseContainer = "List") })
    @RolesAllowed({ Constants.EVI_WS })
    @SecurityRequirement(name = ComandaServerUtils.SECURITY_NAME)
    @Override
    public List<String> llegitUltimesLinies(@PathParam("nomFitxer") @ApiParam("Nom del firxer")
    String nomFitxer, @PathParam("nLinies") @ApiParam("Número de línies a recuperar del firxer")
    Long nLinies) {
        return LogHelper.readLastNLines(LogHelper.getDirectoryLogsFromJbossServerProperties(), nomFitxer, nLinies);
        //return ComandaServerUtils.llegirUltimesLinies(nomFitxer, nLinies);
    }

    /**
     * Retorna una llista amb tots els fitxers que es troben dins la carpeta de logs del servidor de l'aplicació
     *
     * @return successful operation
     */
    @GET
    @Path("/")
    @Produces({ "application/json" })
    @ApiOperation(
            value = "Obtenir el llistat de fitxers de log disponibles",
            notes = "Retorna una llista amb tots els fitxers que es troben dins la carpeta de logs del servidor de l'aplicació",
            tags = { "COMANDA → APP / Logs" })
    @ApiResponses(
            value = { @ApiResponse(
                    code = 200,
                    message = "successful operation",
                    response = FitxerInfo.class,
                    responseContainer = "List") })
    @RolesAllowed({ Constants.EVI_WS })
    @SecurityRequirement(name = ComandaServerUtils.SECURITY_NAME)
    @Override
    public List<FitxerInfo> llistarFitxers() {

        return LogHelper.llistarFitxers(LogHelper.getDirectoryLogsFromJbossServerProperties(), "portafib");
        //return ComandaServerUtils.llistarFitxers();
    }

    /**
     * Descarrega el fitxer de log complet que es troba dins la carpeta de logs del servidor, i que té el nom indicat
     *
     * @param nomFitxer Nom del fitxer
     * @return El fitxer de log
     */
    @GET
    @Path("/{nomFitxer}/directe")
    @Produces({ "application/octet-stream" })
    @ApiOperation(
            value = "Descarregar fitxer de log complet",
            notes = "Descarrega el fitxer de log complet que es troba dins la carpeta de logs del servidor, i que té el nom indicat",
            tags = { "COMANDA → APP / Logs" })
    @ApiResponses(
            value = { @ApiResponse(code = 200, message = "Fitxer descarregat correctament"),
                    @ApiResponse(code = 404, message = "Fitxer no trobat"),
                    @ApiResponse(code = 500, message = "Error intern del servidor") })
    @RolesAllowed({ Constants.EVI_WS })
    @SecurityRequirement(name = ComandaServerUtils.SECURITY_NAME)
    @Override
    public Response descarregarFitxerDirecte(@PathParam("nomFitxer") @ApiParam("Nom del fitxer")
    String nomFitxer) {
        LogFileStream lfs = LogHelper.getFileStreamByNom(LogHelper.getDirectoryLogsFromJbossServerProperties(),
                nomFitxer);

        InputStream is = lfs.getInputStream();
        String contentType = "application/octet-stream";
        String fileName = lfs.getFileName();
        long size = lfs.getSize();

        return Response.ok(is).header("Content-Disposition", "attachment; filename=\"" + fileName + "\"")
                .header("Content-Type", contentType).header("Content-Length", size).build();

    }

}